Renames reg_effects view/viewmodels to dhss
This better matches the file name to the contents. Also, I'll be
adding new views for reg effects and need the filenames for that.
