+++ cegs_portal/search/view_models/__init__.py
@@ -1,5 +1,5 @@
+from .dhss import DHSSearch
 from .experiment import ExperimentSearch
 from .features import FeatureSearch
 from .genes import GeneSearch, IdSearchType, IdType, LocSearchType
-from .reg_effects import DHSSearch
 from .search import Search
