+++ cegs_portal/search/views/__init__.py
@@ -1,6 +1,6 @@
+from .dhss import dhs, dhs_loc
 from .experiment import experiment
 from .features import feature, feature_loc
 from .genes import gene, gene_loc
 from .index import index
-from .reg_effects import dhs, dhs_loc
 from .search import results
